improved security to account crud

+++ app/Livewire/Accounts/ListAccount.php
@@ -24,45 +24,43 @@ class ListAccount extends Component
         $this->dispatch('close');
     }
 
-    public function destroy(Account $account)
+    public function destroy(string $ulid)
     {
+        $account = Account::whereUlid($ulid)->firstOrFail();
         $this->authorize('delete', $account);
-        $account->delete();
 
-        $this->success(__('Record has been deleted successfully'));
-
-        return back();
+        // Cashbook 1 can not be delete
+        if ($account->id !== 1) {
+            $account->delete();
+            $this->success(__('Record has been deleted successfully'));
+        } else {
+            $this->error(__('You can not delete this main account.'));
+        }
     }
 
-    public function forceDelete($id)
+    public function forceDelete(string $ulid)
     {
-        $account = Account::onlyTrashed()->findOrFail($id);
+        $account = Account::onlyTrashed()->whereUlid($ulid)->firstOrFail();
 
         $this->authorize('forceDelete', $account);
         $account->forceDelete();
-
         $this->success(__('Record has been deleted permanently'));
-
-        return back();
     }
 
-    public function restore($id)
+    public function restore(string $ulid)
     {
-        $account = Account::onlyTrashed()->findOrFail($id);
+        $account = Account::onlyTrashed()->whereUlid($ulid)->firstOrFail();
 
         $this->authorize('restore', $account);
         $account->restore();
-
         $this->success(__('Record has been restored successfully'));
-
-        return back();
     }
 
     public function render()
     {
         $this->authorize('viewAny', Account::class);
 
-        $search = $this->search ? '%'.trim($this->search).'%' : null;
+        $search = $this->search ? '%' . trim($this->search) . '%' : null;
         $searchableFields = ['account_no', 'name'];
 
         $accounts = Account::query()
@@ -80,7 +78,7 @@ class ListAccount extends Component
                     $query->withTrashed();
                 }
             })
-            ->latest()
+            ->oldest()
             ->paginate(10);
 
         return view('livewire.accounts.list-account', compact('accounts'))
